refactor(sipher-vault): address B6 hardening code review
- extract MAX_ENCRYPTED_AMOUNT_LEN const (was a bare 64 duplicated across
  the withdraw_private / withdraw_private_sol CPI guards)
- inline the single-use `available` binding in the two withdraw paths
- lift the assertFails test helper into bankrun-helpers.ts (was file-local
  in test 13, re-implementing the shared fail-and-match pattern)
- reconcile SELF-AUDIT.md + DEPLOYMENT.md onto the Anchor 1.0.2 base
  (program version, toolchain table, IDL/blake3 notes, test count 25 -> 39)
- DEPLOYMENT.md: note the downstream sipher IDL/error-code resync required
  at redeploy (removing BalanceLocked renumbers later VaultError codes)
